(PUP-3379) Rename retryaction to retry_action
Renames the code and unit test to be consistent with our naming
conventions.
