+++ lib/puppet/util/retry_action.rb
